+++ src/main/java/de/greluc/sc/sckm/FileHandler.java
@@ -42,7 +42,7 @@ import org.jetbrains.annotations.NotNull;
  *
  * @author Lucas Greuloch (greluc, lucas.greuloch@protonmail.com)
  * @since 1.0.0
- * @version 1.2.1
+ * @version 1.3.0
  */
 @Log4j2
 public class FileHandler {