+++ src/main/java/de/greluc/sc/sckm/data/ChannelType.java
@@ -20,33 +20,11 @@
 
 package de.greluc.sc.sckm.data;
 
-import java.time.ZonedDateTime;
-
-/**
- * Represents an event in which a player is killed during gameplay.
- *
- * <ul>
- *   <li><strong>timestamp</strong>: The date and time when the kill event occurred.
- *   <li><strong>killedPlayer</strong>: The name of the player who was killed.
- *   <li><strong>killer</strong>: The name of the player, NPC, or entity that performed the kill.
- *   <li><strong>weapon</strong>: The weapon or method used to perform the kill.
- *   <li><strong>weaponClass</strong>: The class of the weapon or method used to perform the kill.
- *   <li><strong>damageType</strong>: The type of damage inflicted (e.g., explosive, ballistic).
- *   <li><strong>zone</strong>: The location or area in the game where the kill occurred.
- * </ul>
- *
- * <p>This record provides a detailed representation of a kill event, storing all relevant details
- * for tracking or monitoring purposes.
- *
- * @author Lucas Greuloch (greluc, lucas.greuloch@protonmail.com)
- * @version 1.3.0
- * @since 1.0.0
- */
-public record KillEvent(
-    ZonedDateTime timestamp,
-    String killedPlayer,
-    String killer,
-    String weapon,
-    String weaponClass,
-    String damageType,
-    String zone) {}
+public enum ChannelType {
+  LIVE,
+  PTU,
+  EPTU,
+  HOTFIX,
+  TECH_PREVIEW,
+  CUSTOM
+}