+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - FEATURE: Add class type to ``KillEvent``.
 - FEATURE: Update default scan intervall to be 60 seconds.
 - CHORE: Update dependencies.
+- CHORE: Update persistent storage of settings in registry. **This means, that your settings will be reset to the default values when you start this version for the first time.**
 
 ## [v1.2.0](https://github.com/greluc/SC-Kill-Monitor/releases/tag/v1.2.0)